# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delivery documentation",
        "description": "API for Delivery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:81",
            "description": "Stage Server"
        }
    ],
    "paths": {
        "/api/couriers": {
            "get": {
                "tags": [
                    "Courier"
                ],
                "summary": "Список курьеров",
                "description": "Возращает список курьеров",
                "operationId": "courierIndex",
                "parameters": [
                    {
                        "name": "free",
                        "in": "query",
                        "description": "Только свободные",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список курьеров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Courier"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Courier": {
                "description": "Courier model",
                "properties": {
                    "id": {
                        "description": "Id курьера",
                        "type": "integer"
                    },
                    "credentials": {
                        "description": "ФИО",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Телефон",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Courier"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Токен авторизации",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}